labor_law_qa_project/requirements.txt
requests>=2.31.0
openai>=1.2.3
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.36.2
sentence-transformers>=2.2.2
langchain>=0.1.0
fastapi>=0.104.0
uvicorn>=0.23.0
numpy>=1.24